Add basic functionality for MCU users in mcs-core
Users who join using 'MCU' flag will receive a single video with a composition of all others videos in the room.
MCU sessions are created when the first user MCU user joins the room, and are destroying there are no reamining MCU users in the same room.
MCU sessions are bound to the room object. At this moment we use a single MCU session per room (the default one), but room may have more than one MCU session.
Updated Kurento adapter to be able to handle connection of MCU sessions, creating hubports when needed.
Hubports aren't bound to MCU sessions, but in the media session which wants to connect into (or to receive a connection from) an MCU session.
An small refactoring: in media-session.js, we now await for all the sub 'connect' Promises to be resolved.